fix crate .env file

+++ start.sh
@@ -2,11 +2,23 @@
 ROOT_PATH=$(pwd)
 
 cd laravel
-cp .env.example .env
+if [ -f ".env" ]
+then
+	echo ".env found laravel."
+else
+	echo ".env not found laravel. is being created"
+    cp .env.example .env
+fi
 
 cd $ROOT_PATH
 cd dokcer
-cp .env.example .env
+if [ -f ".env" ]
+then
+	echo ".env found in docker."
+else
+	echo ".env not found in docker. is being created"
+    cp .env.example .env
+fi
 
 docker-compose up --build -d
 docker-compose exec --user devuser php-fpm composer install
